backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Daltro-Oliveira-Vinuto <daltroov777@gmail.com>"]

package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
djangorestframework = "^3.17.1"
psycopg = {extras = ["binary"], version = "^3.3.3"}
django-environ = "^0.13.0"
django-allauth = "^65.15.1"
django-cors-headers = "^4.9.0"
djangorestframework-simplejwt = "^5.5.1"
requests = "^2.33.1"
shell = "^1.0.1"
pillow = "^12.2.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"





